Moved save_geometry to spatial.utils
