+++ boranga/components/occurrence/utils.py
@@ -35,143 +35,6 @@ from boranga.ledger_api_utils import retrieve_email_user
 
 logger = logging.getLogger(__name__)
 
-
-def save_geometry(
-    request,
-    instance,
-    geometry_data,
-    instance_fk_field_name=None,
-    component="occurrence",
-    app_label="boranga",
-):
-    """
-    Save geometry data to the database for an instance of a model.
-    Args:
-        request: The request object
-        instance: The instance of the model to save the geometry for
-        geometry_data: The geometry data to save
-        instance_fk_field_name: The name of the foreign key field on the geometry model
-        component: The component name
-        app_label: The app label
-    """
-
-    instance_model_name = instance._meta.model.__name__
-
-    if not geometry_data:
-        logger.warn(f"No {instance_model_name} geometry to save")
-        return
-
-    InstanceGeometry = apps.get_model(
-        app_label=app_label, model_name=f"{instance_model_name}Geometry"
-    )
-    InstanceGeometrySaveSerializer = getattr(
-        sys.modules[f"boranga.components.{component}.serializers"],
-        f"{instance_model_name}GeometrySaveSerializer",
-    )
-    if instance_fk_field_name is None:
-        instance_fk_field_name = instance_model_name.lower()
-
-    geometry = json.loads(geometry_data)
-    if (
-        0 == len(geometry["features"])
-        and 0
-        == OccurrenceReportGeometry.objects.filter(
-            **{instance_fk_field_name: instance}
-        ).count()
-    ):
-        # No feature to save and no feature to delete
-        logger.warn(f"{instance_model_name} geometry has no features to save or delete")
-        return
-
-    action = request.data.get("action", None)
-
-    geometry_ids = []
-    for feature in geometry.get("features"):
-        supported_geometry_types = ["MultiPolygon", "Polygon", "MultiPoint", "Point"]
-        geometry_type = feature.get("geometry").get("type")
-        # Check if feature is of a supported type, continue if not
-        if geometry_type not in supported_geometry_types:
-            logger.warn(
-                f"{instance_model_name}: {instance} contains a feature that is not a "
-                f"{' or '.join(supported_geometry_types)}: {feature}"
-            )
-            continue
-
-        logger.info(
-            f"Processing {instance_model_name} {instance} geometry feature type: {geometry_type}"
-        )
-
-        geom_4326 = feature_json_to_geosgeometry(feature)
-
-        original_geometry = feature.get("properties", {}).get("original_geometry")
-        srid_original = original_geometry.get("properties", {}).get("srid", 4326)
-        if not srid_original:
-            raise ValidationError(
-                f"Geometry must have an SRID set: {original_geometry.get('coordinates', [])}"
-            )
-
-        if not original_geometry.get("type", None):
-            original_geometry["type"] = geometry_type
-        feature_json = {"type": "Feature", "geometry": original_geometry}
-        geom_original = feature_json_to_geosgeometry(feature_json, srid_original)
-
-        geoms = [(geom_4326, geom_original)]
-
-        for geom in geoms:
-            geometry_data = {
-                f"{instance_fk_field_name}_id": instance.id,
-                "geometry": geom[0],
-                "original_geometry_ewkb": geom[1].ewkb,
-                # TODO: Add intersects condition
-                # "intersects": True,  # probably redunant now that we are not allowing non-intersecting geometries
-            }
-            if feature.get("id"):
-                logger.info(
-                    f"Updating existing {instance_model_name} geometry: {feature.get('id')} for: {instance}"
-                )
-                try:
-                    geometry = InstanceGeometry.objects.get(id=feature.get("id"))
-                except InstanceGeometry.DoesNotExist:
-                    logger.warn(
-                        f"{instance_model_name} geometry does not exist: {feature.get('id')}"
-                    )
-                    continue
-                geometry_data["drawn_by"] = geometry.drawn_by
-                geometry_data["locked"] = (
-                    action in ["submit"]
-                    and geometry.drawn_by == request.user.id
-                    or geometry.locked
-                )
-                serializer = InstanceGeometrySaveSerializer(
-                    geometry, data=geometry_data
-                )
-            else:
-                logger.info(
-                    f"Creating new geometry for {instance_model_name}: {instance}"
-                )
-                geometry_data["drawn_by"] = request.user.id
-                geometry_data["locked"] = action in ["submit"]
-                serializer = InstanceGeometrySaveSerializer(data=geometry_data)
-
-            serializer.is_valid(raise_exception=True)
-            geometry_instance = serializer.save()
-            logger.info(f"Saved {instance_model_name} geometry: {geometry_instance}")
-            geometry_ids.append(geometry_instance.id)
-
-    # Remove any ocr geometries from the db that are no longer in the ocr_geometry that was submitted
-    # Prevent deletion of polygons that are locked after status change (e.g. after submit)
-    # or have been drawn by another user
-    deleted_geometries = (
-        InstanceGeometry.objects.filter(**{instance_fk_field_name: instance})
-        .exclude(Q(id__in=geometry_ids) | Q(locked=True) | ~Q(drawn_by=request.user.id))
-        .delete()
-    )
-    if deleted_geometries[0] > 0:
-        logger.info(
-            f"Deleted {instance_model_name} geometries: {deleted_geometries} for {instance}"
-        )
-
-
 @transaction.atomic
 def ocr_proposal_submit(ocr_proposal, request):
     if not ocr_proposal.can_user_edit:
